created own files for checks
